[16.0][IMP] kmitl_project: absorb budget_project — is_project/project_type on budget.account
Relocate the budget.account project fields (is_project, project_type) from the
now-removed budget_project module into kmitl_project, which already depends on
budget and account_analytic_kmitl. Add an @api.constrains guarding that a code is
not both a project code and a procurement-plan code (ADR-0007), and fix the seed
data to set project_type (project / strategic_project), which the old module
never wrote.
