--- budget_project/models/__init__.py
@@ -1,1 +1,0 @@
-from . import budget_account
