+++ kmitl_project/models/__init__.py
@@ -11,3 +11,4 @@ from . import project_evaluation
 from . import project_expected_outcome
 from . import project_plan
 from . import exception_rule
+from . import budget_account
